docker/dockerizing-django/web/requirements.txt
Django==1.11.4
djangorestframework==3.6.3
django-filter==1.0.4
Markdown==2.6.9
django-jsoneditor==0.0.12
gunicorn==19.3.0
psycopg2==2.6
redis==2.10.3
